requirements.txt
requests
pandas
seaborn
bs4
matplotlib
scikit-learn
tqdm
numpy
regex
pyyaml
